structure de base +fonction de creation du node et graph
